==> 00_covid19_exploration.Rmd <==
---
title: "Exploration of Covid-19 and covariates data"
author: "Marina Espinasse and Sara Rutten"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 2
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: pygments
  pdf_document:
    toc: true
---

```{r, setup, include=FALSE}
require(mosaic)   # Load additional packages here 
knitr::opts_chunk$set(
  tidy=FALSE,     # display code as typed
  size="small",
  message = FALSE,
  warning = FALSE)   # slightly smaller font for code
```

# Libraries

```{r}
library(tidyverse)
library(readxl)
library(here)
library(kableExtra)
library(DT)
library(rebus)
```


# Loading the datasets

Here I will explore the datasets for the thesis: Covid-19 cases registered in each municipality of Belgium, administered vaccines, and the covariates data (N02, 03,PM10, PM2.5, BC)

Data for Covid-19 were downloaded [here](https://epistat.sciensano.be/covid/) and includes two datasets.
The codebook for the data sets is located [here](https://epistat.sciensano.be/COVID19BE_codebook.pdf)

1) Confirmed cases by date and municipality
2) Administered vaccines by week, municipality, age and dose


# Exploring the datasets

## Covid19 cases per municipality dataset

```{r}
cases <- read.csv(here("data", "COVID19BE_CASES_MUNI.csv"))
```

Cases dataset has the following columns: date (of the first symptoms, 2020-2023), NIS municipality code, Dutch and French municipalities' names, Administrative district name in Dutch and French, Belgian Province, Belgian Region, and the number of confirmed cases



Let's first have a quick look at the dataset:

```{r}
glimpse(cases)
```
How many missing data points in each column?

```{r}
map(cases, ~ sum(is.na(.x)))
```

There are 1071 row of data (including all years 2020-2023 and most of the months), where there is no information on municipality. 


How many of such observations without municipality information were in 2021 and 2022?

```{r}
test21 <- cases |> filter(str_detect(DATE, "2021."))
map(test21, ~ sum(is.na(.x)))#365

test22 <- cases |> filter(str_detect(DATE, "2022."))
map(test22, ~ sum(is.na(.x)))#365

```


**Exploring the data without missing information: cases per municipality**

Remove unnecessary columns and rename some. Separate date into year, month and day
```{r}
cases_prep <- cases |> 
  filter(!is.na(NIS5)) |> 
  select(c(NIS5, DATE, PROVINCE, REGION, CASES)) |> 
  rename(mcp_code = NIS5,
         date = DATE,
         province = PROVINCE,
         region = REGION,
         cases = CASES) |> 
  mutate(year = lubridate::year(date),
         month = lubridate::month(date),
         day = lubridate::day(date)
         )
```

Now we can have a look at the frequency of Covid-19 records. How many observations are made per month of the year?
Here I will replace <5 counts of cases with "5", but later we can decide to do it differently.

```{r}
#create a summary table with year, month and mean, max and min cases per month. But first change <5 to 5 to make the column numerical.

cases_summary <- cases_prep |> 
  mutate(cases = ifelse(cases == "<5", 5, cases)) |>
  mutate(cases = as.numeric(cases)) |> 
  group_by(mcp_code, year, month) |> 
  summarize(min = min(cases),
            max=max(cases),
            mean = mean(cases)) 

#the distribution of the mean cases count per municipality
mosaic::fav_stats(cases_summary$mean)
```


**How many records are made per municipality?**

```{r}
library(gt)
```


```{r}
mcp_summary_months <- cases_prep |> 
  group_by(mcp_code, year) |> 
  summarise(n_month = n_distinct(month)) 

  
#kable(mcp_summary_months, "html") %>% kable_styling("striped") %>% scroll_box(width = "100%")
#
datatable(mcp_summary_months)
```

In some municipalities, for instance, Limburg 2021 (73028), cases were registered not every month but only in April, October, November, December.

```{r}
mcp_summary_days <- cases_prep |> 
  group_by(mcp_code, year, month) |> 
  summarise(n_days = n_distinct(day)) 


datatable(mcp_summary_days)
```

Similarly, there will not be registrations of COVID19 every day, in some municipalities and months (like 11004 in June 2020), only 1 record was made. 


### Distribution of <5 values in the COVID19 dataset

How many <5 values are there in total?

```{r}
cases_prep |> 
  filter(cases == "<5") |> 
  nrow()
```

230954 out of 418341, which is 55%.

How many times per month (regardless of a municipality), we observe <5 values in the data set? Here, I will look at 2021 and 2022 year only

**Data for 2021**: there are 153308 observations in total.

```{r}
bias_monthly <- cases_prep |> 
  filter(year %in% c("2021","2022")) |> 
  group_by(year, month) |> 
    summarize(bias_count = sum(cases == "<5"))

breaks1 <- seq(1, 12, by=1)
breaks1y <-seq(2500, 9000, by = 2000)

bias_monthly
```


```{r}
plot1 <- ggplot(data = bias_monthly, aes(x = month, y = bias_count)) +
  geom_col(fill = "darkcyan") +
  scale_x_continuous(breaks = breaks1) +
  scale_y_continuous(breaks = breaks1y) +
  theme_minimal() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12, vjust = +3, face = "bold")) +
  labs(x = "Month", y = "Count of <5 cases")

plot1
```



Only at the end of 2021 the count of <5 was small, but all other months - about the same (except in June 2021).

But what is the proportion of <5 counts per month, dose it differ between months?

```{r}
bias_proportion <- cases_prep |> 
  filter(year %in% c("2021","2022")) |> 
  group_by(year, month) |> 
    summarize(bias_count = sum(cases == "<5"),
              total = n(),
              bias_prop = bias_count/total)

kable(bias_proportion) |> kable_paper("hover", full_width = F)
```



Create a figure of bias proportions

```{r}
plot2 <- ggplot(data = bias_proportion, aes(x = month, y = bias_prop)) +
  geom_col(fill = "darkslategray") +
  scale_x_continuous(breaks = breaks1) +
  scale_y_continuous(breaks = seq(0,1, by = 0.2)) +
  theme_minimal() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 12, vjust = +3, face = "bold")) +
  labs(x = "Month", y = "Proportion of <5 cases")

```



## Administered vaccines dataset

```{r}
vaccine <- read.csv(here("data", "COVID19BE_VACC_MUNI_CUM.csv"))
```

This dataset has the following columns: year_week (combined year and week information), municipality code, age group, dose (1st, 2nd, booster), cumulative number of people receiving the vaccine of a particular dose

Check if there are missing data points in the vaccine dataset: 4317 rows with missing municipality information

```{r}
map(vaccine, ~ sum(is.na(.x)))
```
Let's take a look at rows with missing mcp information

```{r}
vaccine |> 
  filter(is.na(NIS5)) |> 
  slice_head(n=20)
```


* Age groups included are: "18-24" "25-34" "45-54" "55-64" "65-74" "75-84" "85+"   "35-44" "0-17" 

* Type of doses included are: "A"  "C"  "B"  "E"  "E2" "E3"

Here also we have <10 in the number of people receiving the dose. 

Change column names, separate year and week
```{r}
vaccine_prep <- vaccine |> 
  mutate(year = str_extract(YEAR_WEEK, "^[[:digit:]]{2}")) |> 
  mutate(week = str_extract(YEAR_WEEK, "[[:digit:]]{2}$")) |> 
  janitor::clean_names() |> 
  rename(mcp_code = nis5,
         age_group = agegroup,
         cumulative_n_vaccinated = cumul)
```


**How many observations we have per month/ are all weeks included?**

```{r}
vac_weekly <- vaccine_prep |> 
  group_by(year) |> 
  summarize(n_weeks = n_distinct(week))

kable(vac_weekly) |> kable_classic_2(full_width = F)

```

People start to be vaccinated only at the end of 2020, so we have only 1 week with data for that year. The data on 2023 is yet insufficient, likely, we will not use 2023 at all in our analyses.




  
```{r echo=FALSE}
#sessionInfo()  # could use devtools::session_info() if you prefer that

# Documenting R session 

  # * File creation date: `r Sys.Date()`
  # * `r R.version.string`
  # * R version (short form): `r getRversion()`
  # * `mosaic` package version: `r packageVersion("mosaic")`
  # * `inla` package version: `r packageVersion("inla")``
  # * Additional session information

```
