==> 01_data_preparation.Rmd <==
---
title: "Preparation of Covid-19 and Pollution data"
author: "Marina Espinasse and Sara Rutten"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 2
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: pygments
  pdf_document:
    toc: true
---

```{r, setup, include=FALSE}
require(mosaic) # Load additional packages here
knitr::opts_chunk$set(
  tidy = FALSE, # display code as typed
  size = "small",
  message = FALSE,
  warning = FALSE
) # slightly smaller font for code
```


In this file, I will clean and prepare all the datasets for the project: covid19 cases in Belgium (per municipality), administered vaccines data and pollution data.

The goal is to prepare a joint datsets where covid19 cases are combined with the covariates data (pollution and maybe other covariates). This main dataset will be further used in the DLNMs.



```{r, include=FALSE}
library(tidyverse)
library(readxl)
library(here)
library(kableExtra)
library(DT)
library(rebus)
library(data.table)
library(lubridate)
```

# Covid19 cases dataset

```{r}
cases <- read.csv(here("data", "COVID19BE_CASES_MUNI.csv"))
```

I will note all important steps I do for the data cleaning. First, we decided to remove the 1071 rows where there is no information on the municipality or region.

Then, we need to replace the "<5" counts if we random number generated from the `runif` function with minimum of 1 and maximum of 4.

Finally, because records of Covid19 were not made every day (not even every month in some municipalities), we need to make these data points explicit with zero cases of Covid19. In other words, we assume that no registered cases means there were truly no cases in a given municipality at a given month.

```{r}
cases_prep <- cases |>
  filter(!is.na(NIS5)) |>
  select(c(NIS5, DATE, TX_DESCR_NL, PROVINCE, REGION, CASES)) |>
  rename(
    mcp_code = NIS5,
    mcp_name = TX_DESCR_NL,
    date = DATE,
    province = PROVINCE,
    region = REGION,
    cases = CASES
  ) |>
  mutate(
    year = lubridate::year(date),
    month = lubridate::month(date),
    day = lubridate::day(date),
    week = lubridate::week(date),
    cases = ifelse(cases == "<5", NA, cases),
    cases = as.numeric(cases)
  ) 
```


I choose only September 2020 until January 2022 that we will use for analyses

```{r}
days202122 <- data.frame(date = seq(as.Date("2020-09-01"), as.Date("2022-01-31"), by = "+1 day"))

set.seed(110323)

cases_prep2 <- cases_prep |>
  mutate(date = as.Date(date)) |> 
  filter(as.Date("2020-09-01") <= date & date <= as.Date("2022-01-31")) |>
  select(c(mcp_code, date,year, mcp_name, province, region, cases)) |> 
  rowwise() |>
  mutate(cases = if_else(is.na(cases),
    round(runif(1, min = 1, max = 4), 0),
    cases
  ))


cases_bymp <- split(cases_prep2, cases_prep2$mcp_code)
```


Next, I will add calendar dates to each municipality when no Covid19 records were made.

```{r}
cases_bymp_prep <- cases_bymp %>%
  purrr::map(., ~ right_join(.x, days202122, by = "date")) %>%
  purrr::map(., ~ ungroup(.x)) %>%
  purrr::map(., ~ mutate(.x, mcp_code_new = rep(mcp_code[1], 518) )) %>%
  purrr::map(., ~ mutate(.x, mcp_name_new = rep(mcp_name[1], 518))) %>%
  purrr::map(., ~ mutate(.x, province_new = rep(province[1], 518))) %>%
  purrr::map(., ~ mutate(.x, region_new = rep(region[1], 518)))

# check that the function worked
# cases_bymp_prep[["73028"]]
# cases_bymp_prep[["11001"]]
#
# What is the length of each new dataset per municipality? Should be 518 - OK!
# map(cases_bymp_prep, ~ nrow(.x))
```

Finally, I recreate the dataset with all municipalities

```{r}
cases_filled <- do.call("rbind", cases_bymp_prep)

cases_filled_prep <- cases_filled |> 
  select(c(date, mcp_code_new, mcp_name_new,
           province_new, region_new, cases)) |> 
  rename(mcp_code = mcp_code_new,
         mcp_name = mcp_name_new,
         province = province_new,
         region = region_new) |> 
  mutate(
    year = lubridate::year(date),
    month = lubridate::month(date),
    day = lubridate::day(date),
    week = lubridate::week(date))
```

Finally, in this full dataset we can replace NA in cases with zeroes, assuming no record means no cases in a municipality

```{r}
cases_filled_prep2 <- cases_filled_prep |> 
  mutate(cases = ifelse(is.na(cases),
                        0,
                        cases))
```

Save this dataset
```{r}
write.csv(cases_filled_prep2, row.names = F, "./output/covid19_cases_perday_0filled.csv")
```

## Mean cases of Covid19 per municipality

For the additional exploratory figures, I create mean covid19 counts per municipality
```{r}
covid_municipality_mean_total <- cases_filled_prep2 |> 
  group_by(mcp_code, mcp_name, province, region) |> 
  summarize(mean_mun = mean(cases),
            total_mun = sum(cases))
```

```{r}
write.csv(covid_municipality_mean_total, "./output/covid19_municipality_mean_n_total.csv", row.names = F)
```

For the additional exploratory figures, I create mean covid19 counts, for later making maps of monthly mean counts
```{r}
covid_monthly_mean_total <- cases_filled_prep2 |> 
  group_by(mcp_code, mcp_name, province, region,year, month) |> 
  summarize(mean_monthly = mean(cases),
            total_monthly = sum(cases))
```

```{r}
write.csv(covid_monthly_mean_total, "./output/covid19_monthly_mean_n_total.csv", row.names = F)
```


## Aggregate COVID19 cases per week

Since week 53 is just the 31st (and 30 in 2020) of December, I remove this day from the dataset. The same for week 35 in 2020 (only contains September 1) and week 5 in 2022 (only January 29,30 and 31)

```{r}

cases_filled <- read.csv("./output/covid19_cases_perday_0filled.csv")

cases_per_week <- cases_filled |> 
  filter(!week == 53 & !(year==2020 & week==35) & !(year==2022 & week==5))  |> 
 group_by(mcp_code, mcp_name, province, region, year, week) |> 
  summarize(cases_per_week = sum(cases)) |> 
  rename(week_num = week)
          
```
Save this dataset

```{r}
write.csv(cases_per_week, file = "./output/covid19_total_week_v1.csv")
```



## The distribution of Covid19 cases in 2020 and 2021 per municipality

```{r}
dist <- ggplot(data = cases_filled_prep2,
               aes(x=cases)) +
  geom_density(alpha=0.4, fill="seagreen", color="darkgreen") +
  facet_wrap(.~ province) +
  theme_bw() +
  theme(strip.background = element_rect(color = "white", fill = "white"),
        strip.text = element_text(size=12, face = "bold")) +
  labs(y = "Cases density", "COVID19 cases")
  
```

# Pollution data

Here I will prepare pollution data, check if there are any issues and then combine it with covid19 data


```{r}
ozone.2021 <- read.csv(here("data", "fgem_pop_o3_1995_2021.csv"))
no2.2021 <- read.csv(here("data", "fgem_pop_no2_1995_2021.csv"))
bc.2021 <-  read.csv(here("data", "fgem_pop_bc_2011_2021.csv"))
pm25.2021 <-  read.csv(here("data", "fgem_pop_pm25_2008_2021.csv"))
pm10.2021 <-read.csv(here("data", "fgem_pop_pm10_2000_2021.csv"))

ozone.22 <- read.csv(here("data", "fgem_pop_o3_2022_2022.csv"))
no2.22 <- read.csv(here("data", "fgem_pop_no2_2022_2022.csv"))
bc.22 <-  read.csv(here("data", "fgem_pop_bc_2022_2022.csv"))
pm25.22 <-  read.csv(here("data", "fgem_pop_pm25_2022_2022.csv"))
pm10.22 <-read.csv(here("data", "fgem_pop_pm10_2022_2022.csv"))
```

```{r}
ozone <- rbind(ozone.2021,ozone.22%>%filter(!nisnr == "city_name"))
no2   <- rbind(no2.2021,no2.22%>%filter(!nisnr == "city_name"))
bc    <- rbind(bc.2021,bc.22%>%filter(!nisnr == "city_name"))
pm25  <- rbind(pm25.2021,pm25.22%>%filter(!nisnr == "city_name"))
pm10  <- rbind(pm10.2021,pm10.22%>%filter(!nisnr == "city_name"))

```
  
```{r}
city_name <- as.character(ozone[1,])
city_name <- city_name[-1]

ozone_prep <-ozone |> 
  filter(!nisnr == "city_name" ) |> 
  pivot_longer(cols = c(X11001:X93090),
               names_to = "mcp_code",
               values_to = "pollution_ozone" ) |> 
  mutate(city_name = rep(city_name, 10227)) |> 
  mutate(city_name = str_to_sentence(city_name)) |> 
  rename(date = nisnr) |> 
  mutate(year = str_sub(date, start = 1, end = 4),
         month = str_sub(date, start = 5, end = 6),
         day = str_sub(date, start = 7, end = 8)
  ) |> 
  filter(year %in% c("2020", "2021","2022")) |> 
  mutate(mcp_code = str_replace(mcp_code, pattern = "X", replacement = ""),
         date_formatted = lubridate::ymd(date),
         week_num = lubridate::week(as.character(date_formatted))) |> 
  select(-date)
  
```
  

I will do the same changes to the other 4 pollutant datasets

```{r}
bc_city_name <- as.character(bc[1,])
bc_city_name <- bc_city_name[-1]

bc_prep <- bc|> 
  filter(!nisnr == "city_name" ) |> 
  pivot_longer(cols = c(X11001:X93090),
               names_to = "mcp_code",
               values_to = "pollution_bc" ) |> 
  mutate(city_name = rep(bc_city_name, 4383)) |> 
  mutate(city_name = str_to_sentence(city_name)) |> 
  rename(date = nisnr) |> 
  mutate(year = str_sub(date, start = 1, end = 4),
         month = str_sub(date, start = 5, end = 6),
         day = str_sub(date, start = 7, end = 8)
  ) |> 
 filter(year %in% c("2020", "2021","2022")) |> 
  mutate(mcp_code = str_replace(mcp_code, pattern = "X", replacement = ""),
  date_formatted = lubridate::ymd(date),
         week_num = lubridate::week(as.character(date_formatted))) |> 
  select(-date)
```

The total legnth of ozone_prep dataset is 212065, and for bc_prep is also 212065, which is correct (same number of municipalities, months and days of measurement)

```{r}
no2_city_name <- as.character(no2[1,])
no2_city_name <- no2_city_name[-1]


no2_prep <- no2|> 
  filter(!nisnr == "city_name" ) |> 
  pivot_longer(cols = c(X11001:X93090),
               names_to = "mcp_code",
               values_to = "pollution_no2" ) |> 
  mutate(city_name = rep(no2_city_name, 10227)) |> 
  mutate(city_name = str_to_sentence(city_name)) |> 
  rename(date = nisnr) |> 
  mutate(year = str_sub(date, start = 1, end = 4),
         month = str_sub(date, start = 5, end = 6),
         day = str_sub(date, start = 7, end = 8)
  ) |> 
  filter(year %in% c("2020", "2021","2022")) |> 
  mutate(mcp_code = str_replace(mcp_code, pattern = "X", replacement = ""),
         date_formatted = lubridate::ymd(date),
         week_num = lubridate::week(as.character(date_formatted))) |> 
  select(-date)
```

Particulate matter

```{r}
pm10_city_name <- as.character(pm10[1,])
pm10_city_name <- pm10_city_name[-1]


pm10_prep <- pm10|> 
  filter(!nisnr == "city_name" ) |> 
  pivot_longer(cols = c(X11001:X93090),
               names_to = "mcp_code",
               values_to = "pollution_pm10" ) |> 
  mutate(city_name = rep(pm10_city_name, 8401)) |> 
  mutate(city_name = str_to_sentence(city_name)) |> 
  rename(date = nisnr) |> 
  mutate(year = str_sub(date, start = 1, end = 4),
         month = str_sub(date, start = 5, end = 6),
         day = str_sub(date, start = 7, end = 8)
  ) |> 
 filter(year %in% c("2020", "2021", "2022")) |> 
  mutate(mcp_code = str_replace(mcp_code, pattern = "X", replacement = ""),
         date_formatted = lubridate::ymd(date),
         week_num = lubridate::week(as.character(date_formatted))) |> 
  select(-date)
```


```{r}
pm25_city_name <- as.character(pm25[1,])
pm25_city_name <- pm25_city_name[-1]


pm25_prep <- pm25|> 
  filter(!nisnr == "city_name" ) |> 
  pivot_longer(cols = c(X11001:X93090),
               names_to = "mcp_code",
               values_to = "pollution_pm25" ) |> 
  mutate(city_name = rep(pm25_city_name, 5479)) |> 
  mutate(city_name = str_to_sentence(city_name)) |> 
  rename(date = nisnr) |> 
  mutate(year = str_sub(date, start = 1, end = 4),
         month = str_sub(date, start = 5, end = 6),
         day = str_sub(date, start = 7, end = 8)
  ) |> 
  filter(year %in% c("2020", "2021", "2022")) |> 
  mutate(mcp_code = str_replace(mcp_code, pattern = "X", replacement = ""),
         date_formatted = lubridate::ymd(date),
         week_num = lubridate::week(as.character(date_formatted))) |> 
  select(-date)
```

Combine the 5 pollutants in a common dataset

```{r}
pollution <- ozone_prep |> 
  left_join(bc_prep[,c(1,2,4,5,6)], by = c("mcp_code","year", "month", "day")) |> 
  left_join(no2_prep[,c(1,2,4,5,6)], by = c("mcp_code", "year", "month", "day")) |> 
  left_join(pm10_prep[,c(1,2,4,5,6)], by = c("mcp_code","year", "month", "day")) |> 
   left_join(pm25_prep[,c(1,2,4,5,6)], by = c("mcp_code","year", "month", "day")) |> 
  select(c(mcp_code,city_name, year, month, day), everything()) |> 
  mutate(across(matches("pollution"), as.numeric)) |> 
  mutate(day = str_replace(day, pattern = "^0", replacement = ""),
         month = str_replace(month, pattern = "^0", replacement = "")) 
```

Save the dataset of all pollutions per day

```{r}
write.csv(pollution, file = "./output/pollutions_per_day.csv", row.names = F)
```


Check how many days a month pollution measurements were made at each municipality

```{r}
measure_freq <- pollution |> group_by(mcp_code, year, month) |> tally()
```

Each day of the month measurement were taken.

## Check correlations between the pollutants

```{r}
library(ggcorrplot)
library(gridExtra)
library(grid)
```

```{r}
correlations <- round(cor(pollution[,c(6,9:12)]),1)
pvals <- cor_pmat(pollution[,c(6,9:12)], digits = 2)

colnames(correlations) <- c("O3", "BC", "NO2", "PM10", "PM2.5")
rownames(correlations) <- c("O3", "BC", "NO2", "PM10", "PM2.5")


```

```{r}
cors_plot <- ggcorrplot(correlations, hc.order = TRUE, type = "lower",
     outline.col = "white", lab=TRUE, legend.title = "Pearson correlations") 

```

```{r}
ggsave(cors_plot, file = "./figs/correlations_pollutants.pdf")
```
Given the correlation values, we probably have to drop one of the particulate matter pollutants (for instance, we can consider only pm2.5, as a more important pollutant for health risks). Choosing between black carbon and NO2, probably we will keep NO2 as a more important pollutant than black carbon.


## Aggregate pollution per week (mean value over a week)

I remove week 53 from both year 2021 (31st of December) and 2020 (30th and 31st of December)
```{r}
pollution <-read.csv(file = "./output/pollutions_per_day.csv")

pollution_week <- pollution |> 
  filter(!week_num == 53) |> 
  group_by(mcp_code, city_name, year, week_num) |> 
  summarize(pollution_ozone_wk = mean(pollution_ozone),
            pollution_bc_wk = mean(pollution_bc),
            pollution_no2_wk = mean(pollution_no2),
            pollution_pm10_wk = mean(pollution_pm10),
            pollution_pm25_wk = mean(pollution_pm25))
  
```

Save this dataset

```{r}
write.csv(pollution_week, file = "./output/pollution_mean_week.csv", row.names = F)
```




# Prepare vaccination data

```{r}
vaccine <- read.csv(here("data", "COVID19BE_VACC_MUNI_CUM.csv"))
```

For vaccination data, I choose only year 2021 and aggregate vaccines of any type and over all age groups who took the vaccine. There were only 1 week (week 53) with data in vaccination in 2020.
There are 4317 observations where no information on municipality is recorded, I will remove these rows.

```{r}
vaccine_prep <- vaccine |>
  filter(!is.na(NIS5)) |> 
  janitor::clean_names() |> 
  mutate(year = str_extract(year_week, "^[[:digit:]]{2}")) |> 
  filter(!str_detect(year_week, pattern = "^23."))|>
  mutate(week_num = str_extract(year_week, pattern = "[[:digit:]]{2}$")) |> 
  mutate(week_num = str_replace(week_num, pattern = "^0", replacement = "")) |> 
  rename(mcp_code = nis5) 
  
```

Next, I need to replace "<10" cumulative number of people who received the vaccine with a real number.For now, I will do the same as for COVID-19 cases: replace them with a random sample from a uniform from 1 to 9.

There are `sum(str_detect(vaccine_prep$cumul, pattern = "<"))` (or 116396) such values <10 out of total 757079 observations. 

And finally, I count the cumulative number of people as people "fully vaccinated" (i.e. two doses or one dose if only one is necessary) and divide by the population size.

```{r}
pop20_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2020.txt", sep = "|")
pop21_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2021.txt", sep = "|")
pop22_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2022.txt", sep = "|")

pop20_df$agegroup = ifelse(pop20_df$CD_AGE <=17, "0-17", ifelse(pop20_df$CD_AGE>=85, "85+",NA))
for (i in unique(vaccine_prep$agegroup)[-c(7,9)]){
  lb = as.numeric(substr(i,start=1,stop=2))
  ub = as.numeric(substr(i,start=4,stop=5))
  pop20_df$agegroup = ifelse(lb<=pop20_df$CD_AGE & pop20_df$CD_AGE<=ub,i,pop20_df$agegroup)
}

pop21_df$agegroup = ifelse(pop21_df$CD_AGE <=17, "0-17", ifelse(pop21_df$CD_AGE>=85, "85+",NA))
for (i in unique(vaccine_prep$agegroup)[-c(7,9)]){
  lb = as.numeric(substr(i,start=1,stop=2))
  ub = as.numeric(substr(i,start=4,stop=5))
  pop21_df$agegroup = ifelse(lb<=pop21_df$CD_AGE & pop21_df$CD_AGE<=ub,i,pop21_df$agegroup)
}

pop22_df$agegroup = ifelse(pop22_df$CD_AGE <=17, "0-17", ifelse(pop22_df$CD_AGE>=85, "85+",NA))
for (i in unique(vaccine_prep$agegroup)[-c(7,9)]){
  lb = as.numeric(substr(i,start=1,stop=2))
  ub = as.numeric(substr(i,start=4,stop=5))
  pop22_df$agegroup = ifelse(lb<=pop22_df$CD_AGE & pop22_df$CD_AGE<=ub,i,pop22_df$agegroup)
}
```


```{r}
pop20_df = pop20_df %>% group_by(CD_REFNIS, agegroup) %>% summarize(pop_tot = sum(MS_POPULATION)) %>% ungroup()
pop21_df = pop21_df %>% group_by(CD_REFNIS, agegroup) %>% summarize(pop_tot = sum(MS_POPULATION)) %>% ungroup()
pop22_df = pop22_df %>% group_by(CD_REFNIS, agegroup) %>% summarize(pop_tot = sum(MS_POPULATION)) %>% ungroup()

pop_df <- rbind(cbind(pop20_df,year=rep("20",581*9)),
                  cbind(pop21_df,year=rep("21",581*9)),
                  cbind(pop22_df,year=rep("22",581*9)))|> 
  rename(mcp_code = CD_REFNIS) 

vaccine_prep2 = vaccine_prep %>% left_join(pop_df, by=c("mcp_code","year", "agegroup"))
```

```{r}
vaccine_prep3 <- vaccine_prep2 |> 
  mutate(cumul = ifelse(cumul == "<10", NA, cumul)) |> 
  mutate(cumul = ifelse(is.na(cumul),
    round(runif(1, min = 1, max = min(9,pop_tot)), 0),
    cumul
  )
  ) |> 
  mutate(cumul = as.numeric(cumul)) |> 
  filter(dose == "B" | dose == "C") |>
  group_by(mcp_code,year, week_num) |> 
  summarise(cumul_week = sum(cumul))
  
```

Save this dataset

```{r}
write.csv(vaccine_prep3, file = "./output/cumulative_vaccines_week.csv", row.names = F)
```



# Create a complete dataset

I noticed also that year 2021 had 52 full weeks, but the function `lubridate::week()` in R counts the 31st of December as week 53. I will remove this day from Covid dataset


```{r}
covid_df <- read.csv(file = "./output/covid19_total_week_v1.csv")
pollution_df <- read.csv(file = "./output/pollution_mean_week.csv")
vaccine_df <- read.csv(file = "./output/cumulative_vaccines_week.csv")
pop20_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2020.txt", sep = "|")
pop21_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2021.txt", sep = "|")
pop22_df <- read.delim(file = "./data/TF_SOC_POP_STRUCT_2022.txt", sep = "|")
```


First, I aggregate the population over all age groups and both genders.

```{r}
pop20_prep <- pop20_df |> 
  select(c(CD_REFNIS,CD_SEX:MS_POPULATION )) |> 
  group_by(CD_REFNIS) |> 
  summarize(pop_tot = sum(MS_POPULATION)) |> 
  rename(mcp_code = CD_REFNIS) 

pop21_prep <- pop21_df |> 
  select(c(CD_REFNIS,CD_SEX:MS_POPULATION )) |> 
  group_by(CD_REFNIS) |> 
  summarize(pop_tot = sum(MS_POPULATION)) |> 
  rename(mcp_code = CD_REFNIS) 

pop22_prep <- pop22_df |> 
  select(c(CD_REFNIS,CD_SEX:MS_POPULATION )) |> 
  group_by(CD_REFNIS) |> 
  summarize(pop_tot = sum(MS_POPULATION)) |> 
  rename(mcp_code = CD_REFNIS) 

pop_prep <- rbind(cbind(pop20_prep,year=rep(2020,581)),
                  cbind(pop21_prep,year=rep(2021,581)),
                  cbind(pop22_prep,year=rep(2022,581))) %>%
  group_by(mcp_code) %>%
  mutate(pop_avg = mean(pop_tot)) %>% ungroup()

```

I see that in the vaccines dataset, for the first week of 2021 in January, there are only 564 municipalities that had records of vaccinations. In 2020, there are no records of vaccinations at all. In all the over weeks, all 581 municipalities had records of vaccinations.

```{r}
master_dat <- covid_df |>
  left_join(pop_prep, by = c("mcp_code","year")) |>
    left_join(vaccine_df %>% mutate(year=2000+as.numeric(year)), by = c(
    "mcp_code",
    "week_num",
    "year"
  )) |> 
  mutate(cumul_week = ifelse(is.na(cumul_week),0,cumul_week)) |>
  full_join(pollution_df, by = c(
    "mcp_code",
    "year",
    "week_num"
  )) |>
  select(-X) |>
  ungroup() |>
  mutate(vaccine_rate = cumul_week/pop_avg*100) |>
  select(-pop_avg)
 
```



And here are 17 municipalities, for which no vaccination data was recorded in week 1 of year 2021:

```{r}
no_vacdat <- covid_df |>
  left_join(pop_prep, by = c("mcp_code","year")) |>
    left_join(vaccine_df %>% mutate(year=2000+as.numeric(year)), by = c(
    "mcp_code",
    "week_num",
    "year"
  )) |>
  filter(year %in% c("2021")) |> 
  filter(is.na(cumul_week))

no_vacdat$mcp_code
```


# Check collinearity in the complete dataset, for 2021 and 2022

Check also for collinearity of pollution and vaccination data and with cases data

```{r}
master_dat2122 <- master_dat[!((master_dat$year == 2020 & master_dat$week_num<=35) |(master_dat$year == 2022 & master_dat$week_num>=5)),]
correlations_ms <- round(cor(master_dat2122[,c(6, 7, 8, 11:15)], use = "pairwise.complete.obs"),1)
pvals_ms <- cor_pmat(master_dat2122[,c(6, 7, 8, 11:15)], digits = 2)

colnames(correlations_ms) <- c("Cases", "Pop", "Vaccinations",  "O3", "BC", "NO2", "PM10", "PM2.5")
rownames(correlations_ms) <- c("Cases","Pop", "Vaccinations","O3", "BC", "NO2", "PM10", "PM2.5")

cors_plot_ms <- ggcorrplot(correlations_ms, hc.order = TRUE, type = "lower",
     outline.col = "white", lab=TRUE, legend.title = "Pearson correlations") 
```


And save the master dataset and correlations plot

```{r}
write.csv(master_dat, file = "./output/master_dataset.csv", row.names = F)
```

```{r}
ggsave(cors_plot_ms, file = "./figs/correlations_all_vars.pdf")
```
