==> 05_run_models.Rmd <==
---
title: "Running and comparing INLA models"
author: "Marina Espinasse and Sara Rutten"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 2
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: pygments
  pdf_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE)
```


# Call source file with libraries, functions and data

I load the file "04_crossbasis_definition.Rmd", where I created crossbasis variables, response data and random effects covariates, as well as inla function

```{r}
source(knitr::purl("04_crossbasis_definition.rmd", quiet=TRUE)) #use when running for the 1st time
#source("04_crossbasis_definition.R")
```

# Define a baseline model (intercept-only)

The baseline model does not include any of the pollutants or vaccination covariates. It only has a temporal random effect (here RW1) and a spatial random effect for municipality (BYM2 model). 

* Temporal effect: the effect of a week, rw2 model ("replicate" option could be included to say that temporal effects are allowed to be different between the municipalities, but it caused inla to crash) 

* Spatial effect: random spatial effect of municipality, the same for all months 

Constr = TRUE means that we apply constraint to zero sum of the random terms, but anyway, this is a default setting for BYM2, RW1 and RW2.

```{r}
baseformula <- Y ~ f(T1,  model = "rw2", constr = TRUE, scale.model = TRUE, hyper = precision.prior) +
  f(S1, model = "bym2",graph = "output/adjacency.mat", scale.model = TRUE, hyper = precision.prior)
```

```{r}
baseformula_adj_space <- Y ~ f(S1, model = "bym2",graph = "output/adjacency.mat", scale.model = TRUE, hyper = precision.prior)
```

```{r}
baseformula_adj_time <- Y ~ f(T1,  model = "rw2", constr = TRUE, scale.model = TRUE, hyper = precision.prior)
```


# Define formulas for 1-pollutants model

```{r}
formula.ozone <- update.formula(baseformula, ~ . + basis_ozone + basis_vaccine)
formula.no2 <- update.formula(baseformula, ~ . + basis_no2 + basis_vaccine)
formula.pm10 <- update.formula(baseformula, ~ . + basis_pm10 + basis_vaccine)
formula.pm25 <- update.formula(baseformula, ~ . + basis_pm25 + basis_vaccine)
formula.bc <- update.formula(baseformula, ~ . + basis_bc + basis_vaccine)
formula.vaccine <- update.formula(baseformula, ~ . + basis_vaccine)
```


# Run and compare  baseline and 1-pollutant models and vaccine model

List of formulas and labels for the models:

```{r}
formulas <- list(formula.ozone, formula.no2, formula.pm10, formula.pm25, formula.bc, formula.vaccine)
labels <- c("model.ozone", "model.no2", "model.pm10", "model.pm25", "model.bc", "model.vaccine")
```

Run all models and extract DIC, save the model output 

```{r}
models <- lapply(1:length(formulas), 
              function(i) {
                model <- mymodel(formulas[[i]], df)
                save(model, file = paste0("output/models/", labels[i],".RData"))})
```
I also tried a Poisson model, but it was not better than negative binomial models

```{r}
#basemodel_poisson <- mymodel(baseformula, family = "poisson")#326103.5 
```

# Calculate the model comparison diagnostics
log CPO = conditional predictive ordinate
DIC 

```{r}
table0 <- data.table::data.table(Model  = c("ozone", "no2", "pm10", "pm25", "bc", "vaccine"),
                     DIC = NA, logCPO = NA)

for(i in 1:length(formulas))
  {
  load(paste0("output/models/",labels[i],".RData"))
  table0$DIC[i] <- round(model$dic$dic, 0)
  table0$logCPO[i] <- round(-mean(log(model$cpo$cpo), na.rm=T), 4)
 # gcpo = inla.group.cv(model,  groups = groups, strategy = "posterior")
 # table0$logGCPO[i] <- round(-mean(log(gcpo$cv), na.rm=T), 4)
}

# view table
table0
```

During exploration of the data, I have seen that BC is highly correlated with NO2, pm2.5 and pm10


# Multiple covariates models

Try models with multiple covariates. For this model we do not consider, pm25, pm10 and no2, because they are strongly correlated with BC.

```{r}
formula.multmodel1 <- update.formula(baseformula, ~ . +  basis_bc + basis_ozone + basis_vaccine)
formula.multmodel2 <-update.formula(baseformula, ~ . + basis_ozone + basis_no2 + basis_vaccine)
formula.multmodel3 <-update.formula(baseformula, ~ . + basis_pm10 + basis_bc + basis_vaccine)
formula.multmodel4 <-update.formula(baseformula, ~ . + basis_pm25 + basis_bc + basis_vaccine)
formula.multmodel5 <-update.formula(baseformula, ~ . + basis_ozone + basis_pm10 + basis_vaccine)
```

```{r}
formulas_mult <-list(formula.multmodel1,formula.multmodel2,formula.multmodel3, formula.multmodel4, formula.multmodel5)
labels_mult <-c("model.bc_ozone","model.ozone_no2","model.bc_pm10", "model.bc_pm25", "model.ozone_pm10")
```

```{r}
models_mult <- lapply(1:length(formulas_mult), 
              function(i) {
                model <- mymodel(formulas_mult[[i]], df)
                save(model, file = paste0("output/models/", labels_mult[i],".RData"))})
```


These models will be compared by DIC and logscore (minus log of the conditional predictive ordinate). Interpretation of CPO: low cpo for y_i means poor posterior fit for this value of y. cpo's are less then 1, farther away from 1 if the fit is poor. log(cpo) is negative and large if the fit is very poor, but the inverse (-mean(log)) will indicated that small positive value = good fit. It is easier to interpret cpo in this way, to use it alongside with DIC (where smaller=better).

```{r}
table1 <- data.table::data.table(Model  = c("model.bc_ozone","model.ozone_no2","model.bc_pm10", "model.bc_pm25", "model.ozone_pm10"),
                     DIC = NA, logCPO = NA)

for(i in 1:length(formulas_mult))
  {
  load(paste0("output/models/",labels_mult[i],".RData"))
  table1$DIC[i] <- round(model$dic$dic, 0)
  table1$logCPO[i] <- round(-mean(log(model$cpo$cpo), na.rm=T), 4)
  #gcpo = inla.group.cv(model,  groups = groups, strategy = "posterior")
  #table1$logGCPO[i] <- round(-mean(log(gcpo$cv), na.rm=T), 4)
}

# view table
table1
```



# Models adjusted only by space


```{r}
formula.ozone <- update.formula(baseformula_adj_space, ~ . + basis_ozone + basis_vaccine)
formula.no2 <- update.formula(baseformula_adj_space, ~ . + basis_no2 + basis_vaccine)
formula.pm10 <- update.formula(baseformula_adj_space, ~ . + basis_pm10 + basis_vaccine)
formula.pm25 <- update.formula(baseformula_adj_space, ~ . + basis_pm25 + basis_vaccine)
formula.bc <- update.formula(baseformula_adj_space, ~ . + basis_bc + basis_vaccine)
```


List of formulas and labels for the models:

```{r}
formulas <- list(formula.ozone, formula.no2, formula.pm10, formula.pm25, formula.bc)
labels <- c("model.ozone", "model.no2", "model.pm10", "model.pm25", "model.bc")
```

Run all models and extract DIC, save the model output 

```{r}
models <- lapply(1:length(formulas), 
              function(i) {
                model <- mymodel(formulas[[i]], df)
                save(model, file = paste0("output/models/Adjusted Space/", labels[i],".RData"))})
```


# Models adjusted only by time


```{r}
formula.ozone <- update.formula(baseformula_adj_time, ~ . + basis_ozone + basis_vaccine)
formula.no2 <- update.formula(baseformula_adj_time, ~ . + basis_no2 + basis_vaccine)
formula.pm10 <- update.formula(baseformula_adj_time, ~ . + basis_pm10 + basis_vaccine)
formula.pm25 <- update.formula(baseformula_adj_time, ~ . + basis_pm25 + basis_vaccine)
formula.bc <- update.formula(baseformula_adj_time, ~ . + basis_bc + basis_vaccine)
```


List of formulas and labels for the models:

```{r}
formulas <- list(formula.ozone, formula.no2, formula.pm10, formula.pm25, formula.bc)
labels <- c("model.ozone", "model.no2", "model.pm10", "model.pm25", "model.bc")
```

Run all models and extract DIC, save the model output 

```{r}
models <- lapply(1:length(formulas), 
              function(i) {
                model <- mymodel(formulas[[i]], df)
                save(model, file = paste0("output/models/Adjusted Time/", labels[i],".RData"))})
```



# Models unadjusted


```{r}
formula.ozone <-Y ~ basis_ozone + basis_vaccine
formula.no2 <- Y ~  basis_no2 + basis_vaccine
formula.pm10 <- Y ~  basis_pm10 + basis_vaccine
formula.pm25 <- Y ~  + basis_pm25 + basis_vaccine
formula.bc <- Y ~  basis_bc + basis_vaccine
```


List of formulas and labels for the models:

```{r}
formulas <- list(formula.ozone, formula.no2, formula.pm10, formula.pm25, formula.bc)
labels <- c("model.ozone", "model.no2", "model.pm10", "model.pm25", "model.bc")
```

Run all models and extract DIC, save the model output 

```{r}
models <- lapply(1:length(formulas), 
              function(i) {
                model <- mymodel(formulas[[i]], df)
                save(model, file = paste0("output/models/Unadjusted/", labels[i],".RData"))})
```
