==> 06_noint_model_lagplots.Rmd <==
---
title: "Model output and plots"
author: "Marina Espinasse and Sara Rutten"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    toc: true
    toc_depth: 2
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: pygments
  pdf_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE)
```

In this code, I will analyze model output for the selected model (bc_ozone_vaccine). I will plot the random effects, and the plots of lagged effects and main effects of ozone, bc, and vaccine.


# Call the univariate models

Call the selected model - Base + BC 

```{r}
source(knitr::purl("04_crossbasis_definition.rmd", quiet=TRUE))
#source("04_crossbasis_definition.R")

```
## Generate the predictions from DLNM

Here, coef is a list of effects' coefficients for all the basis functions in the model. Below, we extract coefficients only for a particular pollutant, using "grep" function.

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.bc.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix
```

## The  effects of Black Carbon

First, we extract the effect of BC and then estimate predictions. 

```{r}
#find position of the terms associated with BC crossbasis
indt <- grep("basis_bc", model$names.fixed)

# extract predictions from the BC DLNM, center prediction on the 5%  observed BC pollution  (0.06 mg/m3)
predt_bc <- crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
#round(min(data_sort$pollution_bc_wk), 2)
```

# Lagged effect 
```{r}
# lag response for different BC scenarios: median pollution and maximal pollution
pdf("figs/bc_scenario_univariate.pdf", width = 6, height = 6)

# get exposures values
vars <- predt_bc$predvar

# obtain relative risk (RR) fit and upper and lower confidence limits for all exposure variables
rr <- predt_bc$matRRfit
rr.lci <- predt_bc$matRRlow
rr.uci <- predt_bc$matRRhigh

# set relative risk range 
r1 <- min(range(rr, rr.lci, rr.uci))
r2 <- max(range(rr, rr.lci, rr.uci))

# get selected exposure variable positions
md <- which(round(vars, 2) == 0.4)
mx <- which(round(vars, 2) == 1.7)


# define colours
col1 <- brewer.pal(9, "Purples")[4]
tcol1 <- do.call(rgb, c(as.list(col2rgb(col1)), alpha = 255/4, max = 255))

col2 <- brewer.pal(9, "Purples")[9]
tcol2 <- do.call(rgb, c(as.list(col2rgb(col2)), alpha = 255/4, max = 255))


# define x values (lag, by lag)
lagbylag <- seq(0, nlag, 0.5)

# median pollution
plot(lagbylag, rr[md,], col = col1, type = "l", lwd = 1, 
     xlab = "Lag, weeks", ylab = "RR", main = "", 
     ylim = c(0.8,1.8), frame.plot = T, axes = F)
axis(1, at = 0:nlag, labels = 0:nlag)
axis(2)
xx <- c(lagbylag, rev(lagbylag))
yy <- c(rr.lci[md,], rev(rr.uci[md,]))
polygon(xx, yy, col = tcol1, border = tcol1)


# maximal pollution
lines(lagbylag, rr[mx,], col = col2, lwd = 1)
xx <- c(lagbylag, rev(lagbylag))
yy <- c(rr.lci[mx,], rev(rr.uci[mx,]))
polygon(xx, yy, col = tcol2, border = tcol2)
abline(h = 1, lty = 3)

#expression(paste("BC", " " ,"µg"/"m"^3))
legend("topleft",
       legend = c(expression(paste("BC = ","0.4"," ", "µg"/"m"^3)),
                      expression(paste("BC = ","1.7"," ", "µg"/"m"^3))      ),
       col = c(col1, col2), cex = 0.8,
       lwd = 2, lty = 1, bty = "n", 
       y.intersp = 1.2, horiz = F)

#mtext(side = 2, at = r2*1.25, text = "b", las = 2, cex = 1.2, line = 2)

dev.off()
```

# Contour plot
```{r}
# contour plot of exposure-lag-response associations 
pdf("figs/bc_contour_univariate.pdf", width = 8, height = 6)

par(mar=c(5.1, 5, 4.1, 2.1))

y <- predt_bc$predvar
x <- seq(0, nlag, 0.5)
z <- t(predt_bc$matRRfit)

pal <- brewer.pal(9, "Purples")
levels <- pretty(z, 20)
col1 <- colorRampPalette(pal[2:9])
col2<-colorRampPalette(c("white", "grey90"))#grey palette for RR <1
cols_test <- c(col2(sum(levels < 1)), col1(sum(levels > 1)))
cols_test[c(1,2)] <- "#ffffff" #if you want to remove green for RR <1



filled.contour(x,y,z,
               xlab = "Lag, weeks", ylab = expression(paste("BC", " " ,"µg"/"m"^3)), main = "",
               col = cols_test,levels = levels,
               plot.axes = { axis(1, at = 0:nlag, c(0:nlag)) 
                 axis(2)})

#mtext(side = 4, at = max(y), text = "RR", las = 2, cex = 1.2, line = 2)

dev.off()
```


## Plot for cumulative effects

For the following plot, I will also add the quantiles of BC concentrations used for predictions: 
20% = 0.27, 50% =0.44,  95% = 1.2.

```{r}
pdf("figs/bc_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_bc, ptype="overall",  xlab = expression(paste("BC", " " ,"µg"/"m"^3)), ylab="RR", col=6, ylim=c(0,10),
     xlim = c(quantile(data_prep$pollution_bc_wk,0.01),quantile(data_prep$pollution_bc_wk,0.99)))
abline(v=0.27, col="slateblue4", lty = 3, lwd = 2)#20%
abline(v=0.44, col="slateblue4", lty = 3, lwd = 2)#50%
abline(v=1.2, col="slateblue4", lty = 3, lwd = 2)#95%
text(0.27, 50,col="slateblue4", "20%", cex = 1)
text(0.44, 50,col="slateblue4", "50%", cex = 1)
text(1.2, 50,col="slateblue4", "95%", cex = 1)
```

```{r}
BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)
```

# vaccine: Cumulative effect of pollution


```{r}
#find position of the terms associated with BC crossbasis
indt <- grep("basis_vaccine", model$names.fixed)

# extract predictions from the BC DLNM, center prediction on the 5%  observed BC pollution  (0.06 mg/m3)
predt_vaccine <- crosspred(basis_vaccine, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen = quantile(data_prep$vaccine_rate,0.95), cumul = TRUE)
```

```{r}
pdf("figs/vaccine_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_vaccine,col="green4", ptype="overall", xlab="vaccination rate", ylab="RR", main="", ylim = c(0,10)) 
abline(v=0, col="green4", lty = 3, lwd = 2)#20%
abline(v=15.8, col="green4", lty = 3, lwd = 2)#50%
abline(v=83.93, col="green4", lty = 3, lwd = 2)#95%
text(0, 8,col="green4", "20%", cex = 1)
text(15.8, 8,col="green4", "50%", cex = 1)
text(83.93, 8,col="green4", "95%", cex = 1)
```


# Ozone

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.ozone.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix
```

## The  effects of Ozone

First, we extract the effect of O3 and then estimate predictions. 

```{r}
#find position of the terms associated with ozone crossbasis
indt <- grep("basis_ozone", model$names.fixed)

# extract predictions from the ozone DLNM, center prediction on the 5%  observed ozone pollution
predt_ozone <- crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
```


## Plot for cumulative effects

For the following plot, I will also add the quantiles of BC concentrations used for predictions: 
20% = 0.27, 50% =0.44,  95% = 1.2.

```{r}
pdf("figs/ozone_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_ozone,col="grey", ptype="overall", xlab=expression(paste("Ozone", " " ,"µg"/"m"^3)), ylab="RR", main="", ylim = c(0,3),xlim = c(quantile(data_prep$pollution_ozone_wk,0.01),quantile(data_prep$pollution_ozone_wk,0.99))) 
abline(v=31.7, col="grey4", lty = 3, lwd = 2)#20%
abline(v=45.8, col="grey4", lty = 3, lwd = 2)#50%
abline(v=69.9, col="grey4", lty = 3, lwd = 2)#95%
text(31.7, 50,col="grey4", "20%", cex = 1)
text(45.8, 50,col="grey4", "50%", cex = 1)
text(69.9, 50,col="grey4", "95%", cex = 1)
```

```{r}
ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)
```

# NO2

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.no2.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix
```


```{r}
#find position of the terms associated with no2 crossbasis
indt <- grep("basis_no2", model$names.fixed)

# extract predictions from the no2 DLNM, center prediction on the 5% observed no2 pollution
predt_no2 <- crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
```


## Plot for cumulative effects


```{r}
pdf("figs/no2_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_no2,col="grey", ptype="overall", xlab=expression(paste("NO2", " " ,"µg"/"m"^3)), ylab="RR", main="", ylim = c(0,2), xlim = c(quantile(data_prep$pollution_no2_wk,0.01),quantile(data_prep$pollution_no2_wk,0.99))) 
abline(v=6.5, col="grey4", lty = 3, lwd = 2)#20%
abline(v=10.6, col="grey4", lty = 3, lwd = 2)#50%
abline(v=22.5, col="grey4", lty = 3, lwd = 2)#95%
text(6.5, 50,col="grey4", "20%", cex = 1)
text(10.6, 50,col="grey4", "50%", cex = 1)
text(22.5, 50,col="grey4", "95%", cex = 1)
```

```{r}
no2_univariate = crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],at = c(6.5,10.6,22.5),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = data.frame(pred =c(6.5,10.6,22.5), allRR = no2_univariate$allRRfit, allRRlow = no2_univariate$allRRlow, allRRhigh = no2_univariate$allRRhigh)
```

# PM10

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix
```


```{r}
#find position of the terms associated with pm10 crossbasis
indt <- grep("basis_pm10", model$names.fixed)

# extract predictions from the pm10 DLNM, center prediction on the 5%  observed pm10 pollution
predt_pm10 <- crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
```


## Plot for cumulative effects


```{r}
pdf("figs/pm10_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_pm10,col="grey", ptype="overall", xlab=expression(paste("PM10", " " ,"µg"/"m"^3)), ylab="RR", main="", ylim = c(0,2), xlim = c(quantile(data_prep$pollution_pm10_wk,0.01),quantile(data_prep$pollution_pm10_wk,0.99)))  
abline(v=9.1, col="grey4", lty = 3, lwd = 2)#20%
abline(v=15.0, col="grey4", lty = 3, lwd = 2)#50%
abline(v=29.3, col="grey4", lty = 3, lwd = 2)#95%
text(9.1, 50,col="grey4", "20%", cex = 1)
text(15.0, 50,col="grey4", "50%", cex = 1)
text(29.3, 50,col="grey4", "95%", cex = 1)
```

```{r}
pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```

# PM25

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.pm25.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix
```


```{r}
#find position of the terms associated with pm25 crossbasis
indt <- grep("basis_pm25", model$names.fixed)

# extract predictions from the pm25 DLNM, center prediction on the 5%  observed pm25 pollution
predt_pm25 <- crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm25_wk,0.05), 2))
```


## Plot for cumulative effects


```{r}
pdf("figs/pm25_cumul_contour_univariate.pdf", width = 8, height = 6)
plot(predt_pm25,col="grey", ptype="overall", xlab=expression(paste("PM25", " " ,"µg"/"m"^3)), ylab="RR", main="", ylim = c(0,3), xlim = c(quantile(data_prep$pollution_pm25_wk,0.01),quantile(data_prep$pollution_pm25_wk,0.99))) 
abline(v=4.7, col="grey4", lty = 3, lwd = 2)#20%
abline(v=8.5, col="grey4", lty = 3, lwd = 2)#50%
abline(v=18.4, col="grey4", lty = 3, lwd = 2)#95%
text(4.7, 50,col="grey4", "20%", cex = 1)
text(8.5, 50,col="grey4", "50%", cex = 1)
text(18.4, 50,col="grey4", "95%", cex = 1)
```

```{r}
pm25_univariate = crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],at = c(4.7,8.5,18.4),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_sort$pollution_pm25_wk,0.05), 2))
pm25_univariate = data.frame(pred = c(4.7,8.5,18.4), allRR = pm25_univariate$allRRfit, allRRlow = pm25_univariate$allRRlow, allRRhigh = pm25_univariate$allRRhigh)
```






## Adjusted by space

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Space/model.bc.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with BC crossbasis
indt <- grep("basis_bc", model$names.fixed)


predt_bc <- crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))


BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Space/model.ozone.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_ozone", model$names.fixed)

predt_ozone <- crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Space/model.no2.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_no2", model$names.fixed)

predt_no2 <- crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],at = c(6.5,10.6,22.5),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = data.frame(pred =c(6.5,10.6,22.5), allRR = no2_univariate$allRRfit, allRRlow = no2_univariate$allRRlow, allRRhigh = no2_univariate$allRRhigh)
```



```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Space/model.pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_pm10", model$names.fixed)

predt_pm10 <- crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))

pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Space/model.pm25.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix


indt <- grep("basis_pm25", model$names.fixed)


predt_pm25 <- crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm25_wk,0.05), 2))

pm25_univariate = crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],at = c(4.7,8.5,18.4),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_sort$pollution_pm25_wk,0.05), 2))
pm25_univariate = data.frame(pred = c(4.7,8.5,18.4), allRR = pm25_univariate$allRRfit, allRRlow = pm25_univariate$allRRlow, allRRhigh = pm25_univariate$allRRhigh)
```




## Adjusted by time

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Time/model.bc.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix


indt <- grep("basis_bc", model$names.fixed)


predt_bc <- crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))


BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Time/model.ozone.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_ozone", model$names.fixed)

predt_ozone <- crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Time/model.no2.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_no2", model$names.fixed)

predt_no2 <- crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],at = c(6.5,10.6,22.5),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = data.frame(pred =c(6.5,10.6,22.5), allRR = no2_univariate$allRRfit, allRRlow = no2_univariate$allRRlow, allRRhigh = no2_univariate$allRRhigh)
```



```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Time/model.pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_pm10", model$names.fixed)

predt_pm10 <- crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))

pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Adjusted Time/model.pm25.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix


indt <- grep("basis_pm25", model$names.fixed)

predt_pm25 <- crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm25_wk,0.05), 2))

pm25_univariate = crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],at = c(4.7,8.5,18.4),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_sort$pollution_pm25_wk,0.05), 2))
pm25_univariate = data.frame(pred = c(4.7,8.5,18.4), allRR = pm25_univariate$allRRfit, allRRlow = pm25_univariate$allRRlow, allRRhigh = pm25_univariate$allRRhigh)
```



## Unadjusted

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Unadjusted/model.bc.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_bc", model$names.fixed)

predt_bc <- crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))

BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Unadjusted/model.ozone.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_ozone", model$names.fixed)

predt_ozone <- crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)
```

```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Unadjusted/model.no2.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_no2", model$names.fixed)


predt_no2 <- crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],at = c(6.5,10.6,22.5),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = data.frame(pred =c(6.5,10.6,22.5), allRR = no2_univariate$allRRfit, allRRlow = no2_univariate$allRRlow, allRRhigh = no2_univariate$allRRhigh)
```



```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Unadjusted/model.pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

indt <- grep("basis_pm10", model$names.fixed)


predt_pm10 <- crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))

pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/Unadjusted/model.pm25.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix


indt <- grep("basis_pm25", model$names.fixed)


predt_pm25 <- crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm25_wk,0.05), 2))

pm25_univariate = crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],at = c(4.7,8.5,18.4),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_sort$pollution_pm25_wk,0.05), 2))
pm25_univariate = data.frame(pred = c(4.7,8.5,18.4), allRR = pm25_univariate$allRRfit, allRRlow = pm25_univariate$allRRlow, allRRhigh = pm25_univariate$allRRhigh)
```



# Bivariate models



```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.bc_ozone.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with BC crossbasis
indt <- grep("basis_bc", model$names.fixed)

BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)


#find position of the terms associated with ozone crossbasis
indt <- grep("basis_ozone", model$names.fixed)

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)
```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.ozone_no2.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with ozone crossbasis
indt <- grep("basis_ozone", model$names.fixed)

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)

#find position of the terms associated with no2 crossbasis
indt <- grep("basis_no2", model$names.fixed)
no2_univariate = crosspred(basis_no2, coef = coef[indt], vcov=vcov[indt,indt],at = c(6.5,10.6,22.5),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_no2_wk,0.05), 2))
no2_univariate = data.frame(pred =c(6.5,10.6,22.5), allRR = no2_univariate$allRRfit, allRRlow = no2_univariate$allRRlow, allRRhigh = no2_univariate$allRRhigh)

```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.bc_pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with BC crossbasis
indt <- grep("basis_bc", model$names.fixed)

BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)


#find position of the terms associated with pm10 crossbasis
indt <- grep("basis_pm10", model$names.fixed)

pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```


```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.bc_pm25.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with BC crossbasis
indt <- grep("basis_bc", model$names.fixed)

BC_univariate = crosspred(basis_bc, coef = coef[indt], vcov=vcov[indt,indt],at = c(0.27,0.44,1.2),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_bc_wk,0.05),2))
BC_univariate = data.frame(pred = c(0.27,0.44,1.2), allRR = BC_univariate$allRRfit, allRRlow = BC_univariate$allRRlow, allRRhigh = BC_univariate$allRRhigh)


#find position of the terms associated with pm25 crossbasis
indt <- grep("basis_pm25", model$names.fixed)

pm25_univariate = crosspred(basis_pm25, coef = coef[indt], vcov=vcov[indt,indt],at = c(4.7,8.5,18.4),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_sort$pollution_pm25_wk,0.05), 2))
pm25_univariate = data.frame(pred = c(4.7,8.5,18.4), allRR = pm25_univariate$allRRfit, allRRlow = pm25_univariate$allRRlow, allRRhigh = pm25_univariate$allRRhigh)
```



```{r}
#Extract full coef and vcov and create indicators for each term
load("output/models/model.ozone_pm10.RData")
coef <- model$summary.fixed$mean
vcov <- model$misc$lincomb.derived.covariance.matrix

#find position of the terms associated with ozone crossbasis
indt <- grep("basis_ozone", model$names.fixed)

ozone_univariate = crosspred(basis_ozone, coef = coef[indt], vcov=vcov[indt,indt],at = c(31.7,45.8,69.9),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_ozone_wk,0.05),2))
ozone_univariate = data.frame(pred =c(31.7,45.8,69.9), allRR = ozone_univariate$allRRfit, allRRlow = ozone_univariate$allRRlow, allRRhigh = ozone_univariate$allRRhigh)


#find position of the terms associated with pm10 crossbasis
indt <- grep("basis_pm10", model$names.fixed)

pm10_univariate = crosspred(basis_pm10, coef = coef[indt], vcov=vcov[indt,indt],at = c(9.1,15.0,29.3),
                   model.link = "log", bylag = 0.5, cen=round(quantile(data_prep$pollution_pm10_wk,0.05), 2))
pm10_univariate = data.frame(pred = c(9.1,15.0,29.3), allRR = pm10_univariate$allRRfit, allRRlow = pm10_univariate$allRRlow, allRRhigh = pm10_univariate$allRRhigh)
```
